A handful of small changes for code cleanliness
OptionallyEquals -> Swapped to using an if let statement instead of using a nil comparison followed by a force unwrap.

ArrayPlusCleaned -> Swapped to using a filter over a for each for the cleaned function. Clarified the documentation slightly.

CollectionExtensions -> Clarified on documentation slightly.

DictionaryPlusReflection -> Swapped to using a generic over the Any type for its object parameter.